refactor workbench risk attribution route
