+++ src/app/router_registry.py
@@ -102,6 +102,9 @@ from app.routers.workbench_performance_modules import (
     router as workbench_performance_modules_router,
 )
 from app.routers.workbench_risk import router as workbench_risk_router
+from app.routers.workbench_risk_attribution import (
+    router as workbench_risk_attribution_router,
+)
 from app.routers.workbench_risk_concentration import (
     router as workbench_risk_concentration_router,
 )
@@ -162,6 +165,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(workbench_performance_evidence_router)
     app.include_router(workbench_performance_modules_router)
     app.include_router(workbench_risk_router)
+    app.include_router(workbench_risk_attribution_router)
     app.include_router(workbench_risk_concentration_router)
     app.include_router(workbench_risk_drawdown_router)
     app.include_router(workbench_risk_rolling_router)
